fix login logic when user is unactivated

+++ stores/users.js
@@ -17,8 +17,13 @@ export const useUsersStore = defineStore("users", () => {
     const existingUser = users.value.find((user) => user.email == data.email);
     if (existingUser) {
       if (existingUser.password == data.password) {
-        localStorage.setItem("logged-user", JSON.stringify(existingUser));
-        return { success: true, message: "Usuario logueado exitosamente!" };
+        if (existingUser.is_active) {
+          localStorage.setItem("logged-user", JSON.stringify(existingUser));
+          return { success: true, message: "Usuario logueado exitosamente!" };
+        }
+        else {
+          return { success: false, message: "La cuenta asociada a este email esta inactiva." };
+        }
       } else {
         return {
           success: false,
